Improved the responses for the creation commands

+++ HR.LeaveManagement.Application/Features/LeaveAllocations/Handlers/Commands/DeleteLeaveAllocationCommandHandler.cs
@@ -1,6 +1,8 @@
 using AutoMapper;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveAllocations.Requests.Commands;
 using HR.LeaveManagement.Application.Persistence.Contracts;
+using HR.LeaveManagement.Domain;
 using MediatR;
 
 namespace HR.LeaveManagement.Application.Features.LeaveAllocations.Handlers.Commands;
@@ -19,6 +21,9 @@ public class DeleteLeaveAllocationCommandHandler : IRequestHandler<DeleteLeaveAl
     public async Task<Unit> Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
     {
         var leaveAllocation = await leaveAllocationRepository.Get(request.Id);
+
+        if (leaveAllocation == null) throw new NotFoundException(nameof(LeaveAllocation), request.Id);
+
         await leaveAllocationRepository.Delete(leaveAllocation);
         return Unit.Value;
     }
