+++ HR.LeaveManagement.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,6 +1,8 @@
 using AutoMapper;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
 using HR.LeaveManagement.Application.Persistence.Contracts;
+using HR.LeaveManagement.Domain;
 using MediatR;
 
 namespace HR.LeaveManagement.Application.Features.LeaveRequests.Handlers.Commands;
@@ -19,6 +21,9 @@ public class DeleteLeaveRequestCommandHandler : IRequestHandler<DeleteLeaveReque
     public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
     {
         var leaveRequest = await leaveRequestRepository.Get(request.Id);
+
+        if (leaveRequest == null) throw new NotFoundException(nameof(LeaveRequest), request.Id);
+
         await leaveRequestRepository.Delete(leaveRequest);
         return Unit.Value;
     }
